more validation in namespaces

+++ js/namespace-fs.js
@@ -10,6 +10,8 @@
 #                                                               #
 ###############################################################*/
 
+const utils = require('./utils');
+
 class Fs {
     // ************************************
     // readFile
@@ -20,6 +22,12 @@ class Fs {
 
         return new Promise(function (resolve, reject) {
             if (that.objRoot.connected === false || that.objRoot.loggedIn === false) reject(new Error('Not logged in'))
+
+            if (utils.validateTypeOfField(filename, 'string') === false) {
+                reject(new Error('Parameter filename must be a string'))
+
+                return
+            }
 
             // send command
             const opCode = 'fs.readFile'
@@ -49,6 +57,18 @@ class Fs {
 
         return new Promise(function (resolve, reject) {
             if (that.connected === false || that.loggedIn === false) reject(new Error('Not logged in'))
+
+            if (utils.validateTypeOfField(filename, 'string') === false) {
+                reject(new Error('Parameter filename must be a string'))
+
+                return
+            }
+
+            if (utils.validateTypeOfField(data, 'string') === false) {
+                reject(new Error('Parameter data must be a string'))
+
+                return
+            }
 
             // send command
             const opCode = 'fs.writeFile'
@@ -80,6 +100,18 @@ class Fs {
         return new Promise(function (resolve, reject) {
             if (that.connected === false || that.loggedIn === false) reject(new Error('Not logged in'))
 
+            if (utils.validateTypeOfField(filename, 'string') === false) {
+                reject(new Error('Parameter filename must be a string'))
+
+                return
+            }
+
+            if (utils.validateTypeOfField(data, 'string') === false) {
+                reject(new Error('Parameter data must be a string'))
+
+                return
+            }
+
             // send command
             const opCode = 'fs.appendFile'
             that.writer("*3" + RESP3.CRLF +
@@ -110,6 +142,17 @@ class Fs {
         return new Promise(function (resolve, reject) {
             if (that.connected === false || that.loggedIn === false) reject(new Error('Not logged in'))
 
+            if (utils.validateTypeOfField(path, 'string') === false) {
+                reject(new Error('Parameter path must be a string'))
+
+                return
+            }
+
+            if (utils.validateTypeOfField(mask, 'string') === false) {
+                reject(new Error('Parameter mask must be a string'))
+
+                return
+            }
             // send command
             const opCode = 'fs.readDir'
             that.writer("*3" + RESP3.CRLF +
@@ -139,6 +182,18 @@ class Fs {
 
         return new Promise(function (resolve, reject) {
             if (that.connected === false || that.loggedIn === false) reject(new Error('Not logged in'))
+
+            if (utils.validateTypeOfField(path, 'string') === false) {
+                reject(new Error('Parameter path must be a string'))
+
+                return
+            }
+
+            if (utils.validateTypeOfField(mask, 'string') === false) {
+                reject(new Error('Parameter mask must be a string'))
+
+                return
+            }
 
             // send command
             const opCode = 'fs.readTree'
@@ -170,6 +225,12 @@ class Fs {
         return new Promise(function (resolve, reject) {
             if (that.objRoot.connected === false || that.objRoot.loggedIn === false) reject(new Error('Not logged in'))
 
+            if (utils.validateTypeOfField(filename, 'string') === false) {
+                reject(new Error('Parameter filename must be a string'))
+
+                return
+            }
+
             // send command
             const opCode = 'fs.removeFile'
             that.writer("*2" + RESP3.CRLF +
@@ -198,6 +259,18 @@ class Fs {
 
         return new Promise(function (resolve, reject) {
             if (that.objRoot.connected === false || that.objRoot.loggedIn === false) reject(new Error('Not logged in'))
+
+            if (utils.validateTypeOfField(filename, 'string') === false) {
+                reject(new Error('Parameter filename must be a string'))
+
+                return
+            }
+
+            if (utils.validateTypeOfField(newFilename, 'string') === false) {
+                reject(new Error('Parameter newFilename must be a string'))
+
+                return
+            }
 
             // send command
             const opCode = 'fs.renameFile'
@@ -228,6 +301,12 @@ class Fs {
 
         return new Promise(function (resolve, reject) {
             if (that.objRoot.connected === false || that.objRoot.loggedIn === false) reject(new Error('Not logged in'))
+
+            if (utils.validateTypeOfField(filename, 'string') === false) {
+                reject(new Error('Parameter filename must be a string'))
+
+                return
+            }
 
             // send command
             const opCode = 'fs.stat'
@@ -265,6 +344,18 @@ class Fs {
         return new Promise(function (resolve, reject) {
             if (that.objRoot.connected === false || that.objRoot.loggedIn === false) reject(new Error('Not logged in'))
 
+            if (utils.validateTypeOfField(source, 'string') === false) {
+                reject(new Error('Parameter source must be a string'))
+
+                return
+            }
+
+            if (utils.validateTypeOfField(destination, 'string') === false) {
+                reject(new Error('Parameter destination must be a string'))
+
+                return
+            }
+
             // send command
             const opCode = 'fs.copyfile'
             that.writer("*3" + RESP3.CRLF +
@@ -295,6 +386,12 @@ class Fs {
         return new Promise(function (resolve, reject) {
             if (that.objRoot.connected === false || that.objRoot.loggedIn === false) reject(new Error('Not logged in'))
 
+            if (utils.validateTypeOfField(path, 'string') === false) {
+                reject(new Error('Parameter path must be a string'))
+
+                return
+            }
+
             // send command
             const opCode = 'fs.mkdir'
             that.writer("*2" + RESP3.CRLF +
@@ -324,6 +421,12 @@ class Fs {
         return new Promise(function (resolve, reject) {
             if (that.objRoot.connected === false || that.objRoot.loggedIn === false) reject(new Error('Not logged in'))
 
+            if (utils.validateTypeOfField(path, 'string') === false) {
+                reject(new Error('Parameter path must be a string'))
+
+                return
+            }
+
             // send command
             const opCode = 'fs.expandPath'
             that.writer("*2" + RESP3.CRLF +
@@ -352,6 +455,12 @@ class Fs {
 
         return new Promise(function (resolve, reject) {
             if (that.objRoot.connected === false || that.objRoot.loggedIn === false) reject(new Error('Not logged in'))
+
+            if (utils.validateTypeOfField(path, 'string') === false) {
+                reject(new Error('Parameter path must be a string'))
+
+                return
+            }
 
             // send command
             const opCode = 'fs.rmdir'
